Update: Made use of user default storage to save user settings if the app is killed.
How it was done to demo:
Integrated a login screen with credentials "naba" as username and "riaz" as the password.
Upon logging in, the user will be presented with the issue list that was decoded using JSON Decoder.
If the user kills the app and opens it again, the most recent user settings will be used. Meaning, if the user had logged in before killing the app, the user will still be logged in and will not be required to login again.
But if the user deletes and reinstalls the app on the device, user settings will be reset and login will be required.